Mzheldin develop

+++ src/main/java/readerblog/mates/readerblog/services/Implementations/AuthorServiceImpl.java
@@ -8,15 +8,19 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import readerblog.mates.readerblog.entities.Author;
 import readerblog.mates.readerblog.entities.Book;
+import readerblog.mates.readerblog.entities.Category;
+import readerblog.mates.readerblog.entities.Genre;
 import readerblog.mates.readerblog.repositories.AuthorRepository;
 import readerblog.mates.readerblog.services.AuthorService;
 import readerblog.mates.readerblog.services.BookService;
+import readerblog.mates.readerblog.services.CategoryService;
+import readerblog.mates.readerblog.services.GenreService;
+import readerblog.mates.readerblog.utils.Utilities;
 
-import java.math.BigDecimal;
-import java.math.RoundingMode;
 import java.util.ArrayList;
-import java.util.Collections;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 /**
  * @author mzheldin@yandex.ru
@@ -27,6 +31,8 @@ public class AuthorServiceImpl implements AuthorService {
 
     private AuthorRepository authorRepository;
     private BookService bookService;
+    private GenreService genreService;
+    private CategoryService categoryService;
 
     @Autowired
     public void setAuthorRepository(AuthorRepository authorRepository) {
@@ -41,8 +47,9 @@ public class AuthorServiceImpl implements AuthorService {
     @Override
     @Transactional
     public Author save(Author author) {
-        if (author != null){
-            author.setRating(roundingRating(author.getRating()));
+        if (author != null ){
+            if (author.getRating() != null)
+                author.setRating(Utilities.roundingRating(author.getRating()));
             return authorRepository.save(author);
         }
         return null;
@@ -50,8 +57,10 @@ public class AuthorServiceImpl implements AuthorService {
 
     @Override
     @Transactional
-    public Author findOneById(Long id) {
-        return authorRepository.findById(id).get();
+    public Author findOne(Long id) {
+        if (id != null)
+            return authorRepository.findById(id).get();
+        return null;
     }
 
     @Override
@@ -63,7 +72,10 @@ public class AuthorServiceImpl implements AuthorService {
     @Override
     @Transactional
     public List<Author> findByRating(Double ratingMin, Double ratingMax) {
-        return authorRepository.findByRatingBetween(ratingMin, ratingMax);
+        List<Author> authors = new ArrayList<>();
+        if (ratingMin != null && ratingMax != null)
+            return authorRepository.findByRatingBetween(Utilities.roundingRating(ratingMin), Utilities.roundingRating(ratingMax));
+        return authors;
     }
 
     /**
@@ -76,6 +88,7 @@ public class AuthorServiceImpl implements AuthorService {
     @Override
     @Transactional
     public List<Author> findByName(String firstName, String lastName, String patronymicName) {
+        List<Author> authors = new ArrayList<>();
         if (firstName != null && lastName != null && patronymicName != null)
             return authorRepository.findByFirstNameAndLastNameAndPatronymicName(firstName, lastName, patronymicName);
         if (firstName != null && lastName != null)
@@ -90,40 +103,56 @@ public class AuthorServiceImpl implements AuthorService {
             return authorRepository.findByLastName(lastName);
         if (patronymicName != null)
             return authorRepository.findByPatronymicName(patronymicName);
-        return null;
+        return authors;
     }
 
-//    @Override
-//    @Transactional
-//    public List<Author> findByGenre(Long genreId) {
-//        return authorRepository.findAllByBooks(bookService.findByGenreId(genreId));
-//    }
+    @Override
+    public List<Author> findByGenre(Long genreId) {
+        Set<Author> authorSet = new HashSet<>();
+        Genre genre = genreService.findOne(genreId);
+        if (genre != null){
+            List<Book> books = genre.getBooks();
+            for (Book book : books)
+                authorSet.addAll(book.getAuthors());
+        }
+        return new ArrayList<>(authorSet);
+    }
 
-//    @Override
-//    @Transactional
-//    public List<Author> findByCategory(Long categoryId) {
-//        return authorRepository.findAllByBooks(bookService.findByCategoryId(categoryId));
-//    }
+    @Override
+    public List<Author> findByCategory(Long categoryId) {
+        Set<Author> authorSet = new HashSet<>();
+        Category category = categoryService.findOne(categoryId);
+        if (category != null){
+            List<Book> books = category.getBooks();
+            for (Book book : books)
+                authorSet.addAll(book.getAuthors());
+        }
+        return new ArrayList<>(authorSet);
+    }
 
-//    @Override
-//    public List<Long> findIdByGenre(Long genreId) {
-//        List<Long> authorsId = new ArrayList<>();
-//        findByGenre(genreId).forEach(author -> authorsId.add(author.getId()));
-//        return authorsId;
-//    }
+    @Override
+    public List<Long> findIdByGenre(Long genreId) {
+        List<Long> ids = new ArrayList<>();
+        findByGenre(genreId).forEach(author -> ids.add(author.getId()));
+        return ids;
+    }
 
-//    @Override
-//    public List<Long> findIdByCategory(Long categoryId) {
-//        List<Long> authorsId = new ArrayList<>();
-//        findByCategory(categoryId).forEach(author -> authorsId.add(author.getId()));
-//        return authorsId;
-//    }
+    @Override
+    public List<Long> findIdByCategory(Long categoryId) {
+        List<Long> ids = new ArrayList<>();
+        findByCategory(categoryId).forEach(author -> ids.add(author.getId()));
+        return ids;
+    }
 
-//    @Override
-//    @Transactional
-//    public List<Author> findByCategoryAndGenre(Long categoryId, Long genreId) {
-//        return authorRepository.findAllByBooks(bookService.findAllByCategoryIdAndGenreId(categoryId, genreId));
-//    }
+    @Override
+    public List<Author> findByCategoryAndGenre(Long categoryId, Long genreId) {
+        List<Author> authors = new ArrayList<>();
+        List<Author> authorsByGenre = findByGenre(genreId);
+        for (Author author : findByCategory(categoryId))
+            if (authorsByGenre.contains(author))
+                authors.add(author);
+        return authors;
+    }
 
     @Override
     @Transactional
@@ -133,47 +162,58 @@ public class AuthorServiceImpl implements AuthorService {
 
     @Override
     @Transactional
-    public void remove(Long authorId) {
-        authorRepository.removeById(authorId);
-    }
-
-//    @Override
-//    @Transactional
-//    public List<Long> findByBook(Book book) {
-//        List<Long> authorsId = new ArrayList<>();
-//        authorRepository.findAllByBooks(Collections.singletonList(book)).forEach(author -> authorsId.add(author.getId()));
-//        return authorsId;
-//    }
-
-    /**
-     * Округляет double до 1 знака после запятой.
-     * @param rating рейтинг
-     * @return округленный рейтинг
-     */
-    /*TODO Я бы вынес этот метод в пакет utils класс UtilMath или поискал стандартный округлитель*/
-    private Double roundingRating(Double rating){
-        if (rating != null)
-            return BigDecimal.valueOf(rating).setScale(1, RoundingMode.HALF_UP).doubleValue();
+    public Author remove(Long authorId) {
+        if (authorId != null && authorRepository.findById(authorId).isPresent())
+            return authorRepository.removeById(authorId);
         return null;
     }
 
+    @Override
     @Transactional
-    public List<Author> findByLastNameFirstLetter(Character firstLetter){
-        return authorRepository.findByLastNameStartingWith(firstLetter);
+    public List<Author> findByLastNameFirstLetter(String firstLetters){
+        List<Author> authors = new ArrayList<>();
+        if (firstLetters != null)
+            return authorRepository.findAllByLastNameStartingWith(firstLetters);
+        return authors;
     }
 
+    @Override
     @Transactional
-    public List<Author> findAllOrderByLastName(){
+    public List<Author> findByFirstNameFirstLetter(String firstLetters) {
+        List<Author> authors = new ArrayList<>();
+        if (firstLetters != null)
+            return authorRepository.findAllByFirstNameStartingWith(firstLetters);
+        return authors;
+    }
+
+    @Override
+    @Transactional
+    public List<Author> saveAll(List<Author> authors) {
+        List<Author> result = new ArrayList<>();
+        if (authors != null && authors.size() > 0){
+            authors.forEach(author -> author.setRating(Utilities.roundingRating(author.getRating())));
+            return authorRepository.saveAll(authors);
+        }
+        return result;
+    }
+
+    @Override
+    @Transactional
+    public List<Author> findAllByOrderByLastName(){
         return authorRepository.findAllByOrderByLastName();
     }
 
+    @Override
     @Transactional
-    public void updateRating(Long id, Double rating){
-        //TODO на твое усмотрение
-        //Проверка значения не помешает, хотя может именно здесь это будет излишне?
-        if (rating > 10.0 || rating < 0.0) throw new IllegalArgumentException();
-        Author author = findOneById(id);
-        author.setRating(rating);
-        save(author);
+    public Boolean updateRating(Long id, Double rating){
+        if (id != null && rating != null){
+            if (rating > 10.0 || rating < 0.0) throw new IllegalArgumentException();
+            Author author = findOne(id);
+            if (author != null){
+                author.setRating(Utilities.roundingRating(rating));
+                return save(author) != null;
+            }
+        }
+        return false;
     }
 }
